Bugfix - Fix db circular dependency issue - refined controllers

+++ backend/bank/src/main/java/com/hasib/bank/dto/TransactionRequestDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class TransactionRequestDto {
-    private int senderAccountNumber;
-    private int receiverAccountNumber;
+    private long senderAccountNumber;
+    private long receiverAccountNumber;
     private long amount;
 }
